requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
GeoAlchemy2==0.12.5
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
psycopg2==2.9.4
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
Shapely==1.8.5.post1
SQLAlchemy==1.4.41
typing-extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
